Bug Fix Pillow Denial of Service vulnerability (#487)
## Descriptions
An issue was discovered in Pillow prior to 10.0.0. It is a Denial of Service that uncontrollably allocates memory to process a given task, potentially causing a service to crash by having it run out of memory. This occurs for truetype in "ImageFont" when textlength in an "ImageDraw" instance operates on a long text argument.




## Impact
CVE-2023-44271
CWE-400 
CWE-770
`CVSS:3.1/AV:N/AC:L/PR:N/UI:N/S:U/C:N/I:N/A:H`

Co-authored-by: Chris Paxton <cpaxton@fb.com>